+++ jme-opensearch-resource-service/src/main/java/ch/admin/bit/jeap/opensearch/resource/domain/MessagePublisher.java
@@ -2,6 +2,7 @@ package ch.admin.bit.jeap.opensearch.resource.domain;
 
 import ch.admin.bit.jme.opensearch.index.jme.transitdecision.JmeTransitDecisionDataV1;
 import ch.admin.bit.jme.opensearch.index.jme.transitdecision.JmeTransitDecisionDataV2;
+import ch.admin.bit.jme.opensearch.index.jme.transitdecision.JmeTransitDecisionDataV3;
 import ch.admin.bit.jme.opensearch.index.jme.transitdocument.JmeTransitDocumentDataV1;
 
 public interface MessagePublisher {
@@ -11,4 +12,6 @@ public interface MessagePublisher {
     void transitDecisionCreated(JmeTransitDecisionDataV1 transitDecisionDto);
 
     void transitDecisionCreatedV2(JmeTransitDecisionDataV2 transitDecisionDto);
+
+    void transitDecisionCreatedV3(JmeTransitDecisionDataV3 transitDecisionDto);
 }